+++ apps/RawLogViewer/CScanAnimation.cpp
@@ -33,7 +33,7 @@
 #include "xRawLogViewerMain.h"
 
 //(*IdInit(CScanAnimation)
-const long CScanAnimation::ID_RADIOBUTTON1 = wxNewId();
+const long CScanAnimation::ID_LIST_OBS_LABELS = wxNewId();
 const long CScanAnimation::ID_RADIOBUTTON2 = wxNewId();
 const long CScanAnimation::ID_STATICTEXT22 = wxNewId();
 const long CScanAnimation::ID_TEXTCTRL11 = wxNewId();
@@ -72,6 +72,8 @@ CScanAnimation::CScanAnimation(
 {
 	//(*Initialize(CScanAnimation)
 	wxFlexGridSizer* FlexGridSizer4;
+	wxFlexGridSizer* FlexGridSizer4a;
+	wxFlexGridSizer* FlexGridSizer4b;
 	wxFlexGridSizer* FlexGridSizer3;
 	wxFlexGridSizer* FlexGridSizer2;
 	wxFlexGridSizer* FlexGridSizer7;
@@ -82,104 +84,65 @@ CScanAnimation::CScanAnimation(
 		parent, wxID_ANY, _("Animate laser scans"), wxDefaultPosition,
 		wxDefaultSize, wxDEFAULT_DIALOG_STYLE | wxRESIZE_BORDER,
 		_T("wxID_ANY"));
-	FlexGridSizer1 = new wxFlexGridSizer(4, 1, 0, 0);
+	FlexGridSizer1 = new wxFlexGridSizer(3, 1, 0, 0);
 	FlexGridSizer1->AddGrowableCol(0);
-	FlexGridSizer1->AddGrowableRow(2);
-	StaticBoxSizer1 =
-		new wxStaticBoxSizer(wxHORIZONTAL, this, _("Get data from:"));
-	BoxSizer4 = new wxBoxSizer(wxVERTICAL);
-	BoxSizer5 = new wxBoxSizer(wxHORIZONTAL);
-	FlexGridSizer8 = new wxFlexGridSizer(2, 4, 0, 0);
-	FlexGridSizer8->AddGrowableCol(2);
-	rbLoaded = new wxRadioButton(
-		this, ID_RADIOBUTTON1, _("Currently loaded rawlog"), wxDefaultPosition,
-		wxDefaultSize, 0, wxDefaultValidator, _T("ID_RADIOBUTTON1"));
-	rbLoaded->SetValue(true);
-	FlexGridSizer8->Add(
-		rbLoaded, 1, wxALL | wxALIGN_LEFT | wxALIGN_CENTER_VERTICAL, 5);
-	FlexGridSizer8->Add(
-		-1, -1, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
-		5);
-	FlexGridSizer8->Add(
-		-1, -1, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
-		5);
-	FlexGridSizer8->Add(
-		-1, -1, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
-		5);
-	rbFile = new wxRadioButton(
-		this, ID_RADIOBUTTON2, _("Rawlog in file:"), wxDefaultPosition,
-		wxDefaultSize, 0, wxDefaultValidator, _T("ID_RADIOBUTTON2"));
-	FlexGridSizer8->Add(
-		rbFile, 1, wxALL | wxALIGN_LEFT | wxALIGN_CENTER_VERTICAL, 5);
-	StaticText22 = new wxStaticText(
-		this, ID_STATICTEXT22, _("Input file:"), wxDefaultPosition,
-		wxDefaultSize, 0, _T("ID_STATICTEXT22"));
-	FlexGridSizer8->Add(
-		StaticText22, 1, wxALL | wxALIGN_RIGHT | wxALIGN_CENTER_VERTICAL, 5);
-	edFile = new wxTextCtrl(
-		this, ID_TEXTCTRL11, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0,
-		wxDefaultValidator, _T("ID_TEXTCTRL11"));
-	FlexGridSizer8->Add(
-		edFile, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 5);
-	btnPickInput = new wxButton(
-		this, ID_BUTTON5, _("Select..."), wxDefaultPosition, wxDefaultSize, 0,
-		wxDefaultValidator, _T("ID_BUTTON5"));
-	FlexGridSizer8->Add(
-		btnPickInput, 1,
-		wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL, 0);
-	BoxSizer5->Add(
-		FlexGridSizer8, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
-	BoxSizer4->Add(
-		BoxSizer5, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
-	StaticBoxSizer1->Add(
-		BoxSizer4, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
-	FlexGridSizer1->Add(
-		StaticBoxSizer1, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
-	FlexGridSizer4 = new wxFlexGridSizer(1, 7, 0, 0);
-	FlexGridSizer4->AddGrowableCol(5);
+	FlexGridSizer1->AddGrowableRow(1);
+	FlexGridSizer4 = new wxFlexGridSizer(1, 4, 0, 0);
+	FlexGridSizer4->AddGrowableCol(3);
 	FlexGridSizer4->AddGrowableRow(0);
+
+	FlexGridSizer4a = new wxFlexGridSizer(1, 3, 0, 0);
 	btnPlay = new wxButton(
 		this, ID_BUTTON1, _("Start"), wxDefaultPosition, wxDefaultSize, 0,
 		wxDefaultValidator, _T("ID_BUTTON1"));
 	btnPlay->SetDefault();
-	FlexGridSizer4->Add(
+	FlexGridSizer4a->Add(
 		btnPlay, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
 		5);
 	btnStop = new wxButton(
 		this, ID_BUTTON2, _("Stop"), wxDefaultPosition, wxDefaultSize, 0,
 		wxDefaultValidator, _T("ID_BUTTON2"));
 	btnStop->Disable();
-	FlexGridSizer4->Add(
+	FlexGridSizer4a->Add(
 		btnStop, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
 		5);
+	btnClose = new wxButton(
+		this, ID_BUTTON3, _("Close"), wxDefaultPosition, wxDefaultSize, 0,
+		wxDefaultValidator, _T("ID_BUTTON3"));
+	FlexGridSizer4a->Add(
+		btnClose, 1, wxALL | wxALIGN_BOTTOM | wxALIGN_CENTER_HORIZONTAL, 5);
+
+	FlexGridSizer4->Add(
+		FlexGridSizer4a, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP);
+
+	FlexGridSizer4b = new wxFlexGridSizer(1, 2, 0, 0);
+
 	StaticText2 = new wxStaticText(
 		this, ID_STATICTEXT4, _("Animation delay (ms):"), wxDefaultPosition,
 		wxDefaultSize, wxALIGN_CENTRE, _T("ID_STATICTEXT4"));
-	FlexGridSizer4->Add(
+	FlexGridSizer4b->Add(
 		StaticText2, 1,
 		wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL, 5);
 	edDelay = new wxSpinCtrl(
 		this, ID_SPINCTRL2, _T("5"), wxDefaultPosition, wxDefaultSize, 0, 0,
-		1000, 5, _T("ID_SPINCTRL2"));
-	edDelay->SetValue(_T("5"));
-	FlexGridSizer4->Add(
+		1000, 20, _T("ID_SPINCTRL2"));
+	edDelay->SetValue(_T("20"));
+	FlexGridSizer4b->Add(
 		edDelay, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
 		5);
-	cbAllowMix = new wxCheckBox(
-		this, ID_CHECKBOX1, _("Enable mixing of diff. lasers"),
-		wxDefaultPosition, wxDefaultSize, 0, wxDefaultValidator,
-		_T("ID_CHECKBOX1"));
-	cbAllowMix->SetValue(true);
 	FlexGridSizer4->Add(
-		cbAllowMix, 1,
-		wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL, 5);
+		FlexGridSizer4b, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP);
+
+	StaticText3 = new wxStaticText(
+		this, ID_STATICTEXT3, _("Visible sensors:"), wxDefaultPosition,
+		wxDefaultSize);
 	FlexGridSizer4->Add(
-		-1, -1, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 5);
-	btnClose = new wxButton(
-		this, ID_BUTTON3, _("Close"), wxDefaultPosition, wxDefaultSize, 0,
-		wxDefaultValidator, _T("ID_BUTTON3"));
+		StaticText3, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_TOP, 5);
+	lstObsLabels = new wxCheckListBox(
+		this, ID_LIST_OBS_LABELS, wxDefaultPosition, wxDefaultSize, {});
 	FlexGridSizer4->Add(
-		btnClose, 1, wxALL | wxALIGN_BOTTOM | wxALIGN_CENTER_HORIZONTAL, 5);
+		lstObsLabels, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP);
+
 	FlexGridSizer1->Add(
 		FlexGridSizer4, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
 	FlexGridSizer2 = new wxFlexGridSizer(1, 1, 0, 0);
@@ -203,8 +166,8 @@ CScanAnimation::CScanAnimation(
 		slPos, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 5);
 	FlexGridSizer3->Add(
 		FlexGridSizer6, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
-	FlexGridSizer7 = new wxFlexGridSizer(2, 4, 0, 0);
-	FlexGridSizer7->AddGrowableCol(3);
+	FlexGridSizer7 = new wxFlexGridSizer(1, 5, 0, 0);
+	FlexGridSizer7->AddGrowableCol(4);
 	StaticText1 = new wxStaticText(
 		this, ID_STATICTEXT1, _("Rawlog index:"), wxDefaultPosition,
 		wxDefaultSize, 0, _T("ID_STATICTEXT1"));
@@ -224,11 +187,6 @@ CScanAnimation::CScanAnimation(
 	FlexGridSizer7->Add(
 		btnJump, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
 		5);
-	lbNumScans = new wxStaticText(
-		this, ID_STATICTEXT2, _("Number of laser scans: 0"), wxDefaultPosition,
-		wxDefaultSize, 0, _T("ID_STATICTEXT2"));
-	FlexGridSizer7->Add(
-		lbNumScans, 1, wxALL | wxALIGN_LEFT | wxALIGN_CENTER_VERTICAL, 5);
 	cbViewOrtho = new wxCheckBox(
 		this, ID_CHECKBOX2, _("Orthogonal projection"), wxDefaultPosition,
 		wxDefaultSize, 0, wxDefaultValidator, _T("ID_CHECKBOX2"));
@@ -236,17 +194,14 @@ CScanAnimation::CScanAnimation(
 	FlexGridSizer7->Add(
 		cbViewOrtho, 1,
 		wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL, 5);
-	FlexGridSizer7->Add(
-		-1, -1, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
-		5);
-	FlexGridSizer7->Add(
-		-1, -1, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
-		5);
-	lbNumPoints = new wxStaticText(
-		this, ID_STATICTEXT3, _("Number of points: 0"), wxDefaultPosition,
-		wxDefaultSize, 0, _T("ID_STATICTEXT3"));
-	FlexGridSizer7->Add(
-		lbNumPoints, 1, wxALL | wxALIGN_LEFT | wxALIGN_CENTER_VERTICAL, 5);
+
+	edTimestamp = new wxTextCtrl(
+		this, ID_TEXTCTRL11, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0,
+		wxDefaultValidator, _T("ID_TEXTCTRL11"));
+	edTimestamp->SetValue("Timestamp: ");
+	edTimestamp->SetEditable(false);
+	FlexGridSizer7->Add(edTimestamp, 1, wxALL | wxEXPAND, 5);
+
 	FlexGridSizer3->Add(
 		FlexGridSizer7, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
 	FlexGridSizer1->Add(
@@ -256,15 +211,9 @@ CScanAnimation::CScanAnimation(
 	FlexGridSizer1->SetSizeHints(this);
 	Center();
 
-	Bind(
-		wxEVT_RADIOBUTTON, &CScanAnimation::OnrbLoadedSelect, this,
-		ID_RADIOBUTTON1);
-	Bind(wxEVT_RADIOBUTTON, &CScanAnimation::OnrbFile, this, ID_RADIOBUTTON2);
 	Bind(wxEVT_BUTTON, &CScanAnimation::OnbtnPickInputClick, this, ID_BUTTON5);
 	Bind(wxEVT_BUTTON, &CScanAnimation::OnbtnPlayClick, this, ID_BUTTON1);
 	Bind(wxEVT_BUTTON, &CScanAnimation::OnbtnStopClick, this, ID_BUTTON2);
-	Bind(
-		wxEVT_CHECKBOX, &CScanAnimation::OncbAllowMixClick, this, ID_CHECKBOX1);
 	Bind(wxEVT_BUTTON, &CScanAnimation::OnbtnCloseClick, this, ID_BUTTON3);
 	Bind(
 		wxEVT_SLIDER, &CScanAnimation::OnslPosCmdScrollChanged, this,
@@ -279,14 +228,20 @@ CScanAnimation::CScanAnimation(
 	Bind(wxEVT_INIT_DIALOG, &CScanAnimation::OnInit, this, wxID_ANY);
 	//*)
 
+	Bind(
+		wxEVT_CHECKBOX, &CScanAnimation::OncbViewOrthoClick, this,
+		ID_CHECKBOX2);
+
+	Bind(
+		wxEVT_CHECKLISTBOX, &CScanAnimation::OncbViewOrthoClick, this,
+		ID_LIST_OBS_LABELS);
+
 	// Initialize 3D view:
 	auto openGLSceneRef = m_plot3D->getOpenGLSceneRef();
 	openGLSceneRef->insert(mrpt::opengl::CGridPlaneXY::Create(
 		-50, 50, -50, 50, 0 /* z */, 5 /* freq */));
 	openGLSceneRef->insert(mrpt::opengl::stock_objects::CornerXYZSimple(
 		1.0 /*scale*/, 3.0 /*line width*/));
-
-	m_mixlasers = cbAllowMix->GetValue();
 }
 
 CScanAnimation::~CScanAnimation()
@@ -304,6 +259,13 @@ void CScanAnimation::RebuildMaps()
 
 	if (idx >= 0 && idx < (int)rawlog.size())
 	{
+		// Show/hide:
+		for (const auto& kv : m_visibleSensors)
+		{
+			if (m_gl_objects.find(kv.first) == m_gl_objects.end()) continue;
+			m_gl_objects.at(kv.first).obj->setVisibility(kv.second);
+		}
+
 		if (rawlog.getType(idx) == CRawlog::etSensoryFrame)
 		{
 			CSensoryFrame::Ptr sf = rawlog.getAsObservations(idx);
@@ -315,6 +277,7 @@ void CScanAnimation::RebuildMaps()
 			sf->insert(rawlog.getAsObservation(idx));
 			BuildMapAndRefresh(sf.get());
 		}
+		m_plot3D->Refresh();
 	}
 
 	WX_END_TRY
@@ -331,26 +294,30 @@ void CScanAnimation::BuildMapAndRefresh(CSensoryFrame* sf)
 	for (auto& it : *sf)
 		it->load();
 
-	// Mix?
-	if (!m_mixlasers)
-	{
-		// if not, just clear all old objects:
-		for (auto& m_gl_object : m_gl_objects)
+	MRPT_TODO("Fade out old observations by name");
+	// m_gl_objects.clear();
+
+	auto lmbdProcessSensorLabel = [&](const std::string& sNameInMap) {
+		if (m_visibleSensors.find(sNameInMap) == m_visibleSensors.end())
 		{
-			TRenderObject& ro = m_gl_object.second;
-			m_plot3D->getOpenGLSceneRef()->removeObject(
-				ro.obj);  // Remove from the opengl viewport
+			m_visibleSensors[sNameInMap] = true;
+
+			lstObsLabels->AppendString(sNameInMap);
+			lstObsLabels->Check(lstObsLabels->GetCount() - 1, true);
 		}
-		m_gl_objects.clear();
-	}
+	};
 
 	// Insert new scans:
 	mrpt::system::TTimeStamp tim_last = INVALID_TIMESTAMP;
 	bool wereScans = false;
 	for (auto& it : *sf)
 	{
-		const std::string sNameInMap =
-			std::string(it->GetRuntimeClass()->className) + it->sensorLabel;
+		const std::string sNameInMap = it->sensorLabel +
+			std::string(" [Type: ") + it->GetRuntimeClass()->className +
+			std::string("]");
+
+		lmbdProcessSensorLabel(sNameInMap);
+
 		if (IS_CLASS(*it, CObservation2DRangeScan))
 		{
 			CObservation2DRangeScan::Ptr obs =
@@ -560,17 +527,12 @@ void CScanAnimation::BuildMapAndRefresh(CSensoryFrame* sf)
 	}
 
 	// Show timestamp:
-	m_plot3D->getOpenGLSceneRef()->getViewport()->addTextMessage(
-		5, 5,
-		mrpt::format(
-			"Timestamp (UTC): %s (%.06f)",
-			mrpt::system::dateTimeToString(tim_last).c_str(),
-			mrpt::Clock::toDouble(tim_last)),
-		0 /*id*/);
+	edTimestamp->SetValue(mrpt::format(
+		"Timestamp (UTC): %s (%.06f)",
+		mrpt::system::dateTimeToString(tim_last).c_str(),
+		mrpt::Clock::toDouble(tim_last)));
 
-	// Force refresh view:
 	m_plot3D->setCameraProjective(!cbViewOrtho->IsChecked());
-	m_plot3D->Refresh();
 
 	// Post-process: unload 3D observations.
 	for (auto& o : *sf)
@@ -625,8 +587,8 @@ void CScanAnimation::OnbtnPlayClick(wxCommandEvent& event)
 	slPos->Enable();
 }
 
-void CScanAnimation::OnbtnStopClick(wxCommandEvent& event) { m_stop = true; }
-void CScanAnimation::OnbtnCloseClick(wxCommandEvent& event) { Close(); }
+void CScanAnimation::OnbtnStopClick(wxCommandEvent&) { m_stop = true; }
+void CScanAnimation::OnbtnCloseClick(wxCommandEvent&) { Close(); }
 void CScanAnimation::OnslPosCmdScrollChanged(wxCommandEvent&)
 {
 	edIndex->SetValue(slPos->GetValue());
@@ -639,9 +601,9 @@ void CScanAnimation::OnbtnJumpClick(wxCommandEvent& event)
 	RebuildMaps();
 }
 
-void CScanAnimation::OnslPosCmdScroll(wxScrollEvent& event) { RebuildMaps(); }
-void CScanAnimation::OnbtnPickInputClick(wxCommandEvent& event) {}
-void CScanAnimation::OnInit(wxInitDialogEvent& event)
+void CScanAnimation::OnslPosCmdScroll(wxScrollEvent&) { RebuildMaps(); }
+void CScanAnimation::OnbtnPickInputClick(wxCommandEvent&) {}
+void CScanAnimation::OnInit(wxInitDialogEvent&)
 {
 	slPos->SetValue(0);
 	slPos->SetMin(0);
@@ -652,37 +614,19 @@ void CScanAnimation::OnInit(wxInitDialogEvent& event)
 
 	Fit();
 
-	wxCommandEvent dummy;
-
-	OnrbLoadedSelect(dummy);
-}
-
-void CScanAnimation::OnrbLoadedSelect(wxCommandEvent& event)
-{
-	edFile->Enable(false);
-	btnPickInput->Enable(false);
-
 	slPos->Enable(true);
 	edIndex->Enable(true);
 	btnJump->Enable(true);
 }
 
-void CScanAnimation::OnrbFile(wxCommandEvent& event)
+void CScanAnimation::OncbViewOrthoClick(wxCommandEvent&)
 {
-	edFile->Enable(true);
-	btnPickInput->Enable(true);
+	for (size_t i = 0; i < lstObsLabels->GetCount(); i++)
+	{
+		bool visible = lstObsLabels->IsChecked(i);
+		const auto str = lstObsLabels->GetString(i).ToStdString();
+		m_visibleSensors.at(str) = visible;
+	}
 
-	slPos->Enable(false);
-	edIndex->Enable(false);
-	btnJump->Enable(false);
-}
-
-void CScanAnimation::OncbAllowMixClick(wxCommandEvent& event)
-{
-	m_mixlasers = cbAllowMix->GetValue();
-}
-
-void CScanAnimation::OncbViewOrthoClick(wxCommandEvent& event)
-{
 	this->RebuildMaps();
 }